rrr: convert rpc/connection_state to module (rrr.connection_state)
connection_state.cpp is the rrr.connection_state module interface
unit; connection_state.hpp deleted. ConnectionState enum (NEW /
CONNECTING / CONNECTED / DISCONNECTING / DISCONNECTED / FAILED) +
ConnectionStateMachine class with rusty::Cell-backed state and
rusty::Function on_state_change callback. is_valid_transition()
holds the lifecycle transition table.

Includers (2) updated to `import rrr.connection_state;`:
  - src/rrr/rpc/client.hpp, src/rrr/rrr.hpp.

Measurement (rrr + rpcbench, -j32, clean):
  +load_balancer:      —  (incremental only)
  +connection_state:   38.64 s, 175.6 MB PCM

20 modules total; 2.29× baseline.
